map ~text~ like =text=
According to the orgmode docs, =text= is verbatim text, while ~text~ is code.

Until now, I have mapped =text= to <code> and this has to stay.
I was only forced to use ~text~ because "text" contained a =.
So now both =text= and ~text~ map to <code> and nothing maps to <tt> or <pre>.